feat: Add Function to Preview Dataframes
This commit introduces a new function, `previewData`, in `functions.R`. This function allows users to preview the first 1000 rows of a dataframe using `DT::datatable`. The name of the dataframe is used as the caption, which is left-aligned. This addition enhances the data exploration capabilities by providing a quick and easy way to preview dataframes