feat: Add Function to Create Master Nutrient Composition Table
This commit introduces a new function, `createMasterNct`, in functions.R. This function creates a master Nutrient Composition Table (NCT) from a list of NCTs. It selects specified columns from the NCT list and reshapes the data from long to wide format. Each row in the resulting dataframe corresponds to a food item, and each column corresponds to a micronutrient.